requirements.txt
PyPDF2>=3.0.0
Pillow>=11.0.0
img2pdf>=0.6.0
